NAV-25849: Innførere en egen AvslagsperiodeGenerator klasse
