+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vedtak/vedtaksperiode/VedtaksperiodeService.kt
@@ -1,15 +1,12 @@
 package no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode
 
-import no.nav.familie.ks.sak.api.dto.BarnMedOpplysningerDto
 import no.nav.familie.ks.sak.common.BehandlingId
 import no.nav.familie.ks.sak.common.exception.Feil
 import no.nav.familie.ks.sak.common.exception.FunksjonellFeil
-import no.nav.familie.ks.sak.common.util.NullablePeriode
 import no.nav.familie.ks.sak.common.util.TIDENES_ENDE
 import no.nav.familie.ks.sak.common.util.TIDENES_MORGEN
 import no.nav.familie.ks.sak.common.util.erSammeEllerEtter
 import no.nav.familie.ks.sak.common.util.erSenereEnnInneværendeMåned
-import no.nav.familie.ks.sak.common.util.førsteDagIInneværendeMåned
 import no.nav.familie.ks.sak.common.util.sisteDagIMåned
 import no.nav.familie.ks.sak.common.util.storForbokstav
 import no.nav.familie.ks.sak.common.util.tilMånedÅr
@@ -23,11 +20,9 @@ import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingKategori
 import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingRepository
 import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingStatus
 import no.nav.familie.ks.sak.kjerne.behandling.domene.Behandlingsresultat
-import no.nav.familie.ks.sak.kjerne.behandling.steg.registrersøknad.SøknadGrunnlagService
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.domene.Vedtak
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.domene.VedtakRepository
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.refusjonEøs.RefusjonEøsRepository
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.EØSBegrunnelseDB
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.NasjonalEllerFellesBegrunnelseDB
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.UtvidetVedtaksperiodeMedBegrunnelser
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.VedtaksperiodeMedBegrunnelser
@@ -35,7 +30,6 @@ import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.tilVedtaksbegrunnelseFritekst
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.utbetalingsperiodeMedBegrunnelser.UtbetalingsperiodeMedBegrunnelserService
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkår
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.VilkårResultat
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.VilkårsvurderingRepository
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.forskyvning.forskyvVilkårResultater
 import no.nav.familie.ks.sak.kjerne.beregning.AndelTilkjentYtelseMedEndreteUtbetalinger
@@ -43,9 +37,7 @@ import no.nav.familie.ks.sak.kjerne.beregning.AndelerTilkjentYtelseOgEndreteUtbe
 import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.BegrunnelseType
 import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.BegrunnelserForPeriodeContext
 import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.EØSBegrunnelse
-import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.IBegrunnelse
 import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.NasjonalEllerFellesBegrunnelse
-import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.NasjonalEllerFellesBegrunnelse.AVSLAG_UREGISTRERT_BARN
 import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.NasjonalEllerFellesBegrunnelse.REDUKSJON_FRAMTIDIG_OPPHØR_BARNEHAGEPLASS
 import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.tilVedtaksbegrunnelse
 import no.nav.familie.ks.sak.kjerne.eøs.kompetanse.KompetanseService
@@ -67,7 +59,6 @@ class VedtaksperiodeService(
     private val vilkårsvurderingRepository: VilkårsvurderingRepository,
     private val overgangsordningAndelService: OvergangsordningAndelService,
     private val sanityService: SanityService,
-    private val søknadGrunnlagService: SøknadGrunnlagService,
     private val utbetalingsperiodeMedBegrunnelserService: UtbetalingsperiodeMedBegrunnelserService,
     private val andelerTilkjentYtelseOgEndreteUtbetalingerService: AndelerTilkjentYtelseOgEndreteUtbetalingerService,
     private val integrasjonClient: IntegrasjonClient,
@@ -76,6 +67,7 @@ class VedtaksperiodeService(
     private val adopsjonService: AdopsjonService,
     private val endringstidspunktService: EndringstidspunktService,
     private val opphørsperiodeGenerator: OpphørsperiodeGenerator,
+    private val avslagsperiodeGenerator: AvslagsperiodeGenerator,
 ) {
     fun oppdaterVedtaksperiodeMedFritekster(
         vedtaksperiodeId: Long,
@@ -232,7 +224,7 @@ class VedtaksperiodeService(
             }
 
         val utbetalingsperioder = utbetalingsperiodeMedBegrunnelserService.hentUtbetalingsperioder(vedtak)
-        val avslagsperioder = hentAvslagsperioderMedBegrunnelser(vedtak)
+        val avslagsperioder = avslagsperiodeGenerator.genererAvslagsperioder(vedtak)
 
         val vedtaksperioderMedBegrunnelser =
             filtrerUtPerioderBasertPåEndringstidspunkt(
@@ -442,102 +434,6 @@ class VedtaksperiodeService(
         }
     }
 
-    private fun hentAvslagsperioderMedBegrunnelser(vedtak: Vedtak): List<VedtaksperiodeMedBegrunnelser> {
-        val behandling = vedtak.behandling
-        val avslagsperioderFraVilkårsvurdering = hentAvslagsperioderFraVilkårsvurdering(behandling.id, vedtak)
-
-        val avslagsperioderFraEndretUtbetalinger = hentAvslagsperioderFraEndretUtbetalinger(behandling.id, vedtak)
-
-        val uregistrerteBarn =
-            if (behandling.erSøknad()) {
-                søknadGrunnlagService.hentAktiv(behandlingId = behandling.id).hentUregistrerteBarn()
-            } else {
-                emptyList()
-            }
-
-        return if (uregistrerteBarn.isNotEmpty()) {
-            leggTilAvslagsbegrunnelseForUregistrertBarn(
-                avslagsperioder = avslagsperioderFraVilkårsvurdering + avslagsperioderFraEndretUtbetalinger,
-                vedtak = vedtak,
-                uregistrerteBarn = uregistrerteBarn,
-            )
-        } else {
-            avslagsperioderFraVilkårsvurdering + avslagsperioderFraEndretUtbetalinger
-        }
-    }
-
-    private fun hentAvslagsperioderFraEndretUtbetalinger(
-        behandlingId: Long,
-        vedtak: Vedtak,
-    ): List<VedtaksperiodeMedBegrunnelser> {
-        val endreteUtbetalinger =
-            andelerTilkjentYtelseOgEndreteUtbetalingerService.finnEndreteUtbetalingerMedAndelerTilkjentYtelse(
-                behandlingId,
-            )
-
-        val periodegrupperteAvslagEndreteUtbetalinger = endreteUtbetalinger.filter { it.erEksplisittAvslagPåSøknad == true }.groupBy { NullablePeriode(it.fom?.toLocalDate(), it.tom?.toLocalDate()) }
-
-        val avslagsperioder =
-            periodegrupperteAvslagEndreteUtbetalinger
-                .map { (fellesPeriode, endretUtbetalinger) ->
-
-                    val avslagsbegrunnelser =
-                        endretUtbetalinger
-                            .map { it.vedtaksbegrunnelser }
-                            .flatten()
-                            .toSet()
-                            .toList()
-
-                    lagVedtaksPeriodeMedBegrunnelser(vedtak, fellesPeriode, avslagsbegrunnelser)
-                }.toMutableList()
-
-        return avslagsperioder
-    }
-
-    private fun hentAvslagsperioderFraVilkårsvurdering(
-        behandlingId: Long,
-        vedtak: Vedtak,
-    ): MutableList<VedtaksperiodeMedBegrunnelser> {
-        val vilkårsvurdering =
-            vilkårsvurderingRepository.finnAktivForBehandling(behandlingId = behandlingId) ?: throw Feil(
-                "Fant ikke vilkårsvurdering for behandling $behandlingId ved generering av avslagsperioder",
-            )
-
-        val andelerTilkjentYtelse = andelerTilkjentYtelseOgEndreteUtbetalingerService.finnAndelerTilkjentYtelseMedEndreteUtbetalinger(behandlingId)
-
-        val periodegrupperteAvslagsvilkår: Map<NullablePeriode, List<VilkårResultat>> =
-            vilkårsvurdering.personResultater
-                .flatMap { it.vilkårResultater }
-                .filter { it.erEksplisittAvslagPåSøknad == true }
-                .groupBy {
-                    val finnesAndelForPersonSomSlutterSammeMånedSomFomIVilkårResultat =
-                        andelerTilkjentYtelse.any { andel ->
-                            andel.aktør == it.personResultat?.aktør && andel.stønadTom == it.periodeFom?.toYearMonth()
-                        }
-                    if (finnesAndelForPersonSomSlutterSammeMånedSomFomIVilkårResultat) {
-                        NullablePeriode(it.periodeFom?.plusMonths(1), it.periodeTom)
-                    } else {
-                        NullablePeriode(it.periodeFom, it.periodeTom)
-                    }
-                }
-
-        val avslagsperioder =
-            periodegrupperteAvslagsvilkår
-                .map { (fellesPeriode, vilkårResultater) ->
-
-                    val avslagsbegrunnelser =
-                        vilkårResultater
-                            .map { it.begrunnelser }
-                            .flatten()
-                            .toSet()
-                            .toList()
-
-                    lagVedtaksPeriodeMedBegrunnelser(vedtak, fellesPeriode, avslagsbegrunnelser)
-                }.toMutableList()
-
-        return avslagsperioder
-    }
-
     fun beskrivPerioderMedRefusjonEøs(
         behandling: Behandling,
         avklart: Boolean,
@@ -572,69 +468,6 @@ class VedtaksperiodeService(
                 }
             }.toSet()
             .takeIf { it.isNotEmpty() }
-    }
-
-    private fun lagVedtaksPeriodeMedBegrunnelser(
-        vedtak: Vedtak,
-        periode: NullablePeriode,
-        avslagsbegrunnelser: List<IBegrunnelse>,
-    ): VedtaksperiodeMedBegrunnelser =
-        VedtaksperiodeMedBegrunnelser(
-            vedtak = vedtak,
-            fom = periode.fom?.førsteDagIInneværendeMåned(),
-            tom = periode.tom?.sisteDagIMåned(),
-            type = Vedtaksperiodetype.AVSLAG,
-        ).apply {
-            begrunnelser.addAll(
-                avslagsbegrunnelser.filterIsInstance<NasjonalEllerFellesBegrunnelse>().map { begrunnelse ->
-                    NasjonalEllerFellesBegrunnelseDB(
-                        vedtaksperiodeMedBegrunnelser = this,
-                        nasjonalEllerFellesBegrunnelse = begrunnelse,
-                    )
-                },
-            )
-
-            eøsBegrunnelser.addAll(
-                avslagsbegrunnelser.filterIsInstance<EØSBegrunnelse>().map { eøsBegrunnelse ->
-                    EØSBegrunnelseDB(vedtaksperiodeMedBegrunnelser = this, begrunnelse = eøsBegrunnelse)
-                },
-            )
-            vedtaksperiodeHentOgPersisterService.lagre(this)
-        }
-
-    private fun leggTilAvslagsbegrunnelseForUregistrertBarn(
-        avslagsperioder: List<VedtaksperiodeMedBegrunnelser>,
-        vedtak: Vedtak,
-        uregistrerteBarn: List<BarnMedOpplysningerDto>,
-    ): List<VedtaksperiodeMedBegrunnelser> {
-        val avslagsperioderMedTomPeriode =
-            if (avslagsperioder.none { it.fom == null && it.tom == null }) {
-                avslagsperioder +
-                    VedtaksperiodeMedBegrunnelser(
-                        vedtak = vedtak,
-                        fom = null,
-                        tom = null,
-                        type = Vedtaksperiodetype.AVSLAG,
-                    )
-            } else {
-                avslagsperioder
-            }
-
-        return avslagsperioderMedTomPeriode
-            .map {
-                if (it.fom == null && it.tom == null && uregistrerteBarn.isNotEmpty()) {
-                    it.apply {
-                        begrunnelser.add(
-                            NasjonalEllerFellesBegrunnelseDB(
-                                vedtaksperiodeMedBegrunnelser = this,
-                                nasjonalEllerFellesBegrunnelse = AVSLAG_UREGISTRERT_BARN,
-                            ),
-                        )
-                    }
-                } else {
-                    it
-                }
-            }.toList()
     }
 }
